+++ packages/smooth_app/lib/cards/data_cards/image_upload_card.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
@@ -26,9 +25,11 @@ class ImageUploadCard extends StatefulWidget {
 }
 
 class _ImageUploadCardState extends State<ImageUploadCard> {
-  ImageProvider? _imageProvider; // Normal size image to display in carousel
-  ImageProvider?
-      _imageFullProvider; // Full resolution image to display in image page
+  /// Normal size image to display in carousel
+  ImageProvider? _imageProvider;
+
+  /// Full resolution image to display in image page
+  ImageProvider? _imageFullProvider;
 
   Future<void> _getImage() async {
     final File? croppedImageFile =
@@ -46,34 +47,15 @@ class _ImageUploadCardState extends State<ImageUploadCard> {
       if (!mounted) {
         return;
       }
-      final bool isUploaded = await uploadCapturedPicture(
+      await uploadCapturedPicture(
         context,
-        barcode: widget.product
-            .barcode!, //Probably throws an error, but this is not a big problem when we got a product without a barcode
+        barcode: widget.product.barcode!,
         imageField: widget.productImageData.imageField,
         imageUri: croppedImageFile.uri,
       );
-      croppedImageFile.delete();
+
       if (!mounted) {
         return;
-      }
-      if (isUploaded) {
-        if (widget.productImageData.imageField == ImageField.OTHER) {
-          final AppLocalizations appLocalizations =
-              AppLocalizations.of(context);
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              content: Text(appLocalizations.other_photo_uploaded),
-              duration: const Duration(seconds: 3),
-              action: SnackBarAction(
-                label: appLocalizations.more_photos,
-                onPressed: _getImage,
-              ),
-            ),
-          );
-        } else {
-          await widget.onUpload(context);
-        }
       }
     }
   }
@@ -132,9 +114,7 @@ class _ImageUploadCardState extends State<ImageUploadCard> {
             context,
             MaterialPageRoute<bool>(
               builder: (BuildContext context) => ProductImageGalleryView(
-                productImageData: widget.productImageData,
-                allProductImagesData: widget.allProductImagesData,
-                title: widget.productImageData.title,
+                imagesData: widget.allProductImagesData,
                 barcode: widget.product.barcode,
               ),
             ),