+++ packages/smooth_app/lib/pages/onboarding/onboarding_flow_navigator.dart
@@ -192,6 +192,7 @@ class OnboardingFlowNavigator {
       child: Builder(
         builder: (BuildContext context) => SmoothScaffold(
           body: widget,
+          brightness: Brightness.dark,
         ),
       ),
     );